Renames syllable counter
